[项目名称：企小微租户系统][提交人：汤密持][提交说明：人力预算计算修改]

+++ qixiaowei-service-api/qixiaowei-operate-cloud-api/src/main/java/net/qixiaowei/operate/cloud/api/dto/employee/EmployeeBudgetDTO.java
@@ -92,7 +92,7 @@ public class EmployeeBudgetDTO {
     /**
      * 年度平均人数
      */
-    private Integer annualAverageNum;
+    private BigDecimal annualAverageNum;
     /**
     * 删除标记:0未删除;1已删除
     */
